:broom: cleanup tests and split model module
