+++ src/graia/ariadne/entry/__init__.py
@@ -27,20 +27,7 @@ from ..exception import MessageTooLong as MessageTooLong
 from ..exception import UnknownError as UnknownError
 from ..exception import UnknownTarget as UnknownTarget
 from ..exception import UnVerifiedSession as UnVerifiedSession
-from ..model import BotMessage as BotMessage
-from ..model import Client as Client
-from ..model import DownloadInfo as DownloadInfo
-from ..model import FileInfo as FileInfo
-from ..model import Friend as Friend
-from ..model import Group as Group
-from ..model import GroupConfig as GroupConfig
-from ..model import LogConfig as LogConfig
-from ..model import Member as Member
-from ..model import MemberInfo as MemberInfo
-from ..model import MemberPerm as MemberPerm
-from ..model import MiraiSession as MiraiSession
-from ..model import Profile as Profile
-from ..model import Stranger as Stranger
+from ..model import *
 from ..util import cpu_bound as cpu_bound
 from ..util import io_bound as io_bound
 from ..util.cooldown import CoolDown as CoolDown
