+++ src/test_old/minimal.py
@@ -1,9 +1,10 @@
 from graia.ariadne.app import Ariadne
+from graia.ariadne.entry import config
 from graia.ariadne.message.chain import MessageChain
 from graia.ariadne.message.element import Plain
-from graia.ariadne.model import Friend, MiraiSession
+from graia.ariadne.model import Friend
 
-app = Ariadne(MiraiSession(host="http://localhost:8080", verify_key="ServiceVerifyKey", account=123456789))
+app = Ariadne(config(123456789, "ServiceVerifyKey"))
 
 
 @app.broadcast.receiver("FriendMessage")
